+++ CHANGES.rst
@@ -47,6 +47,10 @@ refpix
 
 - Update cal_step, add suffix and add to the exposure pipeline [#890]
 
+resample
+--------
+- Implement resampling step. [#787]
+
 
 0.12.0 (2023-08-18)
 ===================